+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 ### Added
 
-- Add decompiled tab (this replaces the decompile toolbar action)
+- New decompiled tab (this replaces the decompile toolbar action)
 - Add Kotlin metadata overview to the structure view
-- Add list relates class files (e.g., nested classes for lambdas) to the structure view
-- Add option to build the whole project if a class file is outdated or missing
+- List relates class files (e.g., nested classes for lambdas) to the structure view
+- New option to build the whole project if a class file is outdated or missing
 - Add option to re-compile the class file in the "Re-Parse Class File" action if the class file is outdated or missing
+- Add source file information to the structure view
 
 ### Changed